UPDATE NAME PACKAGES
I update the name ofthe package project for unihogar, so everything is changed
