+++ ProyectoPadre/ProyectoPersistencia/src/main/java/co/edu/uniquindio/unihogar/entities/RatingPK.java
@@ -1,4 +1,4 @@
-package co.edu.uniquindio.project.entities;
+package co.edu.uniquindio.unihogar.entities;
 
 import java.io.Serializable;
